+++ backend/dto/requests/tasks/get_tasks_request.go
@@ -1,4 +1,4 @@
-package requests
+package tasks
 
 type GetTasksRequest struct {
 	Limit int `query:"limit" validate:"required,min=1,max=100"`
